fix compiling warning

+++ sourceCode/Network/TcpClient.cpp
@@ -162,7 +162,7 @@ TcpResult TcpClient::send(const Serialize::WriteBuffer& buffer)
         restart();
         return TcpResult::Failed;
     }
-    else if (sendBytes != buffer.getDataSize())
+    else if (static_cast<unsigned int>(sendBytes) != buffer.getDataSize())
     {
         TRACE_NOTICE("send message fragmantation! plan send:" << buffer.getDataSize() << ", actuall send:" << sendBytes);
         return TcpResult::Success;
